requirements.txt
streamlit
Pillow
rembg
pandas
opencv-python